refactor: split SessionBoundaryLogging tests into focused modules (#818)
- Split 533-line integration test into 3 focused files
- Create shared test helpers to reduce duplication
- sessionStart.test.tsx (148 lines): Session start workflow tests
- sessionEnd.test.tsx (143 lines): Session end workflow tests
- lifecycle.test.tsx (160 lines): Complete lifecycle tests
- testHelpers.tsx (122 lines): Shared mock factories

Each file now under 300-line limit with single responsibility

Related to #818
